+++ data/input_urls.csv
@@ -62,3 +62,4 @@ Acuode Colecalciferol 50.000UI Polvo,Farmacias Ahumada,https://www.farmaciasahum
 Hormogel,Farmacia Meki,https://farmaciameki.cl/producto/hormogel-pump-05-mg-silesia-0f21
 Gestel 200mg,Farmacia Meki,https://farmaciameki.cl/producto/gestel-cenabast-200-mg-exeltis-110a
 Lenzetto,Profar,https://www.profar.cl/lenzetto-103312/p
+Vitamina D3 (Colecalciferol),Knop Laboratorios,https://www.farmaciasknop.com/products/vitamina-d3-800-ui
